Rename to use new package in Kernel
